pyproject.toml
[tool.poetry]
name = "signup_api"
version = "0.1.0"
description = "An asynchronous Fast API service for signups and Telegram integration"
authors = ["mbrav <mbrav@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "*"
fastapi = "*"
fastapi-pagination = {extras = ["sqlalchemy"], version = "*"}
uvicorn = "*"
SQLAlchemy = "*"
PyMySQL = "*"
python-jose = {extras = ["cryptography"], version = "*"}
passlib = {extras = ["argon2"], version = "*"}
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.9.0"}


[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
autopep8 = "*"
requests = "*"

[tool.pytest.ini_options]
minversion = "6.2"
testpaths = [
    "tests/",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
addopts = "-vv -p no:cacheprovider"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
